+++ packages/headless/src/controllers/commerce/search-box/headless-search-box.ts
@@ -1,8 +1,12 @@
 import {CommerceEngine} from '../../../app/commerce-engine/commerce-engine';
 import {configuration} from '../../../app/common-reducers';
 import {stateKey} from '../../../app/state-key';
+import {
+  registerQuerySetQuery,
+  updateQuerySetQuery,
+} from '../../../features/commerce/query-set/query-set-actions';
 import {fetchQuerySuggestions} from '../../../features/commerce/query-suggest/query-suggest-actions';
-import {UpdateQueryActionCreatorPayload} from '../../../features/commerce/query/query-actions';
+import {UpdateQueryPayload} from '../../../features/commerce/query/query-actions';
 import {queryReducer as commerceQuery} from '../../../features/commerce/query/query-slice';
 import {executeSearch} from '../../../features/commerce/search/search-actions';
 import {
@@ -10,10 +14,6 @@ import {
   prepareForSearchWithQuery,
 } from '../../../features/commerce/search/search-actions';
 import {commerceSearchReducer as commerceSearch} from '../../../features/commerce/search/search-slice';
-import {
-  registerQuerySetQuery,
-  updateQuerySetQuery,
-} from '../../../features/query-set/query-set-actions';
 import {querySetReducer as querySet} from '../../../features/query-set/query-set-slice';
 import {
   clearQuerySuggest,
@@ -110,11 +110,11 @@ export function buildSearchBox(
   const getValue = () => getState().querySet[options.id];
 
   const performSearch = async () => {
-    const queryOptions: UpdateQueryActionCreatorPayload &
-      PrepareForSearchWithQueryOptions = {
-      query: getValue(),
-      clearFilters: options.clearFilters,
-    };
+    const queryOptions: UpdateQueryPayload & PrepareForSearchWithQueryOptions =
+      {
+        query: getValue(),
+        clearFilters: options.clearFilters,
+      };
 
     dispatch(prepareForSearchWithQuery(queryOptions));
     dispatch(executeSearch());