+++ packages/headless/src/features/commerce/recommendations/recommendations-actions.ts
@@ -1,11 +1,14 @@
-import {StringValue} from '@coveo/bueno';
+import {RecordValue, StringValue} from '@coveo/bueno';
+import {Relay} from '@coveo/relay';
 import {createAction, createAsyncThunk} from '@reduxjs/toolkit';
 import {
   AsyncThunkCommerceOptions,
   isErrorResponse,
 } from '../../../api/commerce/commerce-api-client';
+import {ChildProduct} from '../../../api/commerce/common/product';
 import {CommerceRecommendationsRequest} from '../../../api/commerce/recommendations/recommendations-request';
 import {RecommendationsCommerceSuccessResponse} from '../../../api/commerce/recommendations/recommendations-response';
+import {NavigatorContext} from '../../../app/navigatorContextProvider';
 import {RecommendationsSection} from '../../../state/state-sections';
 import {validatePayload} from '../../../utils/validate-payload';
 import {
@@ -28,12 +31,19 @@ export interface QueryRecommendationsCommerceAPIThunkReturn {
 export type StateNeededByFetchRecommendations = StateNeededByQueryCommerceAPI &
   RecommendationsSection;
 
-const buildRecommendationCommerceAPIRequest = async (
+const buildRecommendationCommerceAPIRequest = (
   slotId: string,
   state: StateNeededByFetchRecommendations,
+  relay: Relay,
+  navigatorContext: NavigatorContext,
   productId?: string
-): Promise<CommerceRecommendationsRequest> => {
-  const commerceAPIRequest = await buildBaseCommerceAPIRequest(state, slotId);
+): CommerceRecommendationsRequest => {
+  const commerceAPIRequest = buildBaseCommerceAPIRequest(
+    state,
+    relay,
+    navigatorContext,
+    slotId
+  );
   return {
     ...commerceAPIRequest,
     context: {
@@ -45,7 +55,7 @@ const buildRecommendationCommerceAPIRequest = async (
   };
 };
 
-export interface FetchRecommendationsActionCreatorPayload {
+export interface FetchRecommendationsPayload {
   /**
    * The unique identifier of the recommendations slot (e.g., `b953ab2e-022b-4de4-903f-68b2c0682942`).
    */
@@ -55,15 +65,20 @@ export interface FetchRecommendationsActionCreatorPayload {
 
 export const fetchRecommendations = createAsyncThunk<
   QueryRecommendationsCommerceAPIThunkReturn,
-  FetchRecommendationsActionCreatorPayload,
+  FetchRecommendationsPayload,
   AsyncThunkCommerceOptions<StateNeededByFetchRecommendations>
 >(
   'commerce/recommendations/fetch',
-  async (payload, {getState, rejectWithValue, extra: {apiClient}}) => {
+  async (
+    payload,
+    {getState, rejectWithValue, extra: {apiClient, relay, navigatorContext}}
+  ) => {
     const {slotId, productId} = payload;
-    const request = await buildRecommendationCommerceAPIRequest(
+    const request = buildRecommendationCommerceAPIRequest(
       slotId,
       getState(),
+      relay,
+      navigatorContext,
       productId
     );
     const fetched = await apiClient.getRecommendations(request);
@@ -78,13 +93,18 @@ export const fetchRecommendations = createAsyncThunk<
   }
 );
 
+export type FetchMoreRecommendationsPayload = FetchRecommendationsPayload;
+
 export const fetchMoreRecommendations = createAsyncThunk<
   QueryRecommendationsCommerceAPIThunkReturn | null,
-  FetchRecommendationsActionCreatorPayload,
+  FetchMoreRecommendationsPayload,
   AsyncThunkCommerceOptions<StateNeededByFetchRecommendations>
 >(
   'commerce/recommendations/fetchMore',
-  async (payload, {getState, rejectWithValue, extra: {apiClient}}) => {
+  async (
+    payload,
+    {getState, rejectWithValue, extra: {apiClient, relay, navigatorContext}}
+  ) => {
     const slotId = payload.slotId;
     const state = getState();
     const moreRecommendationsAvailable = moreRecommendationsAvailableSelector(
@@ -100,7 +120,12 @@ export const fetchMoreRecommendations = createAsyncThunk<
     const nextPageToRequest = numberOfProducts / perPage;
 
     const request = {
-      ...(await buildRecommendationCommerceAPIRequest(slotId, state)),
+      ...buildRecommendationCommerceAPIRequest(
+        slotId,
+        state,
+        relay,
+        navigatorContext
+      ),
       page: nextPageToRequest,
     };
     const fetched = await apiClient.getRecommendations(request);
@@ -119,21 +144,26 @@ export interface SlotIdPayload {
   slotId: string;
 }
 
+export type RegisterRecommendationsSlotPayload = SlotIdPayload;
+
 export const registerRecommendationsSlot = createAction(
   'commerce/recommendations/registerSlot',
-  (payload: SlotIdPayload) =>
+  (payload: RegisterRecommendationsSlotPayload) =>
     validatePayload(payload, recommendationsSlotDefinition)
 );
 
 export interface PromoteChildToParentActionCreatorPayload
   extends SlotIdPayload {
-  childPermanentId: string;
-  parentPermanentId: string;
+  child: ChildProduct;
 }
 
 export const promoteChildToParentDefinition = {
-  childPermanentId: new StringValue({required: true}),
-  parentPermanentId: new StringValue({required: true}),
+  child: new RecordValue({
+    options: {required: true},
+    values: {
+      permanentid: new StringValue({required: true}),
+    },
+  }),
   ...recommendationsSlotDefinition,
 };