+++ packages/atomic/CHANGELOG.md
@@ -1,3 +1,21 @@
+# 2.69.0 (2024-06-26)
+
+### Bug Fixes
+
+- **atomic:** add query-summary E2E tests ([#4102](https://github.com/coveo/ui-kit/issues/4102)) ([6f73da7](https://github.com/coveo/ui-kit/commits/6f73da703b5d948b713680d187f03da1d207976a))
+- **atomic:** adjust product-list placeholder ([#4108](https://github.com/coveo/ui-kit/issues/4108)) ([9b28869](https://github.com/coveo/ui-kit/commits/9b288691bd05dac40ac87e244d0c892379469535))
+- **atomic:** clear recent searches on keyboard enter keypress ([#4072](https://github.com/coveo/ui-kit/issues/4072)) ([9ab7ba5](https://github.com/coveo/ui-kit/commits/9ab7ba5f43197e7691051daf45dc955f8a87d8af))
+- **atomic:** escape searchQuery in regex ([#4116](https://github.com/coveo/ui-kit/issues/4116)) ([dbd33a1](https://github.com/coveo/ui-kit/commits/dbd33a18b8eab4ff3c34aa2cd2069f7c043c7f96))
+- **atomic:** fix atomic-commerce-search-box needlessly requesting suggestions when input is disabled ([#4096](https://github.com/coveo/ui-kit/issues/4096)) ([49e9415](https://github.com/coveo/ui-kit/commits/49e9415a601924996d87f1062ea87fed231dcb36))
+- **atomic:** resolve result to full collection in context of folded result list ([#4119](https://github.com/coveo/ui-kit/issues/4119)) ([360b604](https://github.com/coveo/ui-kit/commits/360b604badd96f21c20701e55a109cb8a0766a75))
+- **headless:** update product interface to reflect how the commerce api behaves ([#4104](https://github.com/coveo/ui-kit/issues/4104)) ([5e514c5](https://github.com/coveo/ui-kit/commits/5e514c5bf642e709c6bd897f693ec42174026c24))
+
+### Features
+
+- **atomic:** add alt text field on atomic-result-image ([#4056](https://github.com/coveo/ui-kit/issues/4056)) ([3d69e97](https://github.com/coveo/ui-kit/commits/3d69e976a380c8f71b9a6720be19920cc7fe4447))
+- **atomic:** add identifier on atomic commerce generated facets ([#4112](https://github.com/coveo/ui-kit/issues/4112)) ([b2a4406](https://github.com/coveo/ui-kit/commits/b2a4406f6f94d6d724811ce430484b02eb9d48f5))
+- **commerce:** make summary a sub-controller ([#4082](https://github.com/coveo/ui-kit/issues/4082)) ([cfcc9d9](https://github.com/coveo/ui-kit/commits/cfcc9d9b2557345a0a96e271a0dba9cb549812b4))
+
 ## 2.68.1 (2024-06-19)
 
 ### Bug Fixes